refactor(components): refactor InboxItems file structure
