+++ packages/frontend-design-poc/src/components/InboxItem/index.ts
@@ -1,1 +1,2 @@
 export { InboxItem } from "./InboxItem.tsx";
+export { InboxItems } from "../InboxItem/InboxItems.tsx";
