--- packages/frontend-design-poc/src/components/InboxItems/index.ts
@@ -1,1 +1,0 @@
-export { InboxItems } from "./InboxItems.tsx";
